requirements.txt
streamlit
fastapi
uvicorn
python-multipart
python-dotenv
pydantic
pydantic-settings
langchain
langchain-openai
langchain-community
langchain-core
chromadb
sentence-transformers
pytesseract
pdf2image
pillow
python-magic
python-magic-bin
beautifulsoup4
lxml
openpyxl
pandas
numpy
scikit-learn
tqdm
nltk
fitz
requests

# Development dependencies
pytest
black
isort
flake8

# Database dependencies
psycopg2-binary
psycopg2

